Refactor AnalyzeTry()
